bcProjKso/requirements.txt
mysqlclient
django-cors-headers
